requirements.txt
luma.oled
timeloop
requests
beautifulsoup4